Move resources to the resources package
