+++ js/components/common/ja-spinner/JASpinnerStyle.js
@@ -1,5 +1,5 @@
 import { StyleSheet } from 'react-native';
-import { PRIMARY_THEME_COLOR } from '../../common/colors';
+import { PRIMARY_THEME_COLOR } from '../../../resources/colors';
 
 const JASpinnerColor = PRIMARY_THEME_COLOR;
