+++ js/components/common/ja-tagline/style.js
@@ -1,5 +1,5 @@
 import { StyleSheet } from 'react-native';
-import { PRIMARY_THEME_COLOR } from '../../common/colors';
+import { PRIMARY_THEME_COLOR } from '../../../resources/colors';
 
 const styles = StyleSheet.create({
   center: {
