+++ js/components/common/preview-job-card/previewHeaderView.js
@@ -3,7 +3,7 @@ import { View } from 'react-native';
 
 import CardImageHeader from '../../common/card-image-header/cardImageHeader';
 import SimpleCardBody from '../../common/simple-card-body/simpleCardBody';
-import { imageProp } from '../../common/propTypes';
+import { imageProp } from '../../../resources/propTypes';
 
 const PreviewHeaderView = ({ title, cover, icon, subtitle }) =>
   <View>
