+++ js/components/developer/personal-info-screen/index.js
@@ -3,7 +3,7 @@ import { View, StyleSheet } from 'react-native';
 import { connect } from 'react-redux';
 import { Container, Content, Form, Col, Row, Thumbnail, Card, CardItem, Button, Text } from 'native-base';
 import styles from './style';
-import GlobalStyle from '../../common/globalStyle';
+import GlobalStyle from '../../../resources/globalStyle';
 import EmailInput from '../../common/email-input';
 import PhoneInput from '../../common/numeric-input';
 import PasswordInput from '../../common/password-input';
