+++ js/components/developer/personal-info-screen/style.js
@@ -1,5 +1,5 @@
 import { StyleSheet } from 'react-native';
-import { BLACK } from '../../common/colors';
+import { BLACK } from '../../../resources/colors';
 
 const styles = StyleSheet.create({
   form: {
