+++ js/resources/colors.js
