+++ js/resources/globalStyle.js
