+++ js/tabBarStyles.js
@@ -1,5 +1,5 @@
 import { StyleSheet } from 'react-native';
-import { PRIMARY_THEME_COLOR, WHITE, GRAY } from './components/common/colors';
+import { PRIMARY_THEME_COLOR, WHITE, GRAY } from './resources/colors';
 
 const ACTIVE_TINT_COLOR = PRIMARY_THEME_COLOR;
 const INACTIVE_TINT_COLOR = GRAY;
